refactored to Stats component
