+++ src/layout/content/ContentStyles.tsx
@@ -24,41 +24,6 @@ export const MainContainer = styled.main`
 	}
 `
 
-export const StatsWrapper = styled.div`
-	width: 100%;
-	overflow: hidden;
-	margin-bottom: 40px;
-	@media (max-width: ${breakpoints.md}px) {
-		overflow: auto;
-	}
-`
-
-export const StatsOverlayWrapper = styled(Row)`
-	display: flex;
-	width: 100%;
-	flex-wrap: wrap;
-	.ant-col:not(:last-child) {
-		padding-right: 16px;
-	}
-	@media (max-width: ${breakpoints.xl}px) and (min-width: ${breakpoints.md}px) {
-		.ant-col {
-			padding-right: 0 !important;
-			&:nth-child(1),
-			&:nth-child(2) {
-				padding-bottom: 16px;
-			}
-			&:nth-child(1),
-			&:nth-child(3) {
-				padding-right: 16px !important;
-			}
-		}
-	}
-	@media (max-width: ${breakpoints.md}px) {
-		justify-content: space-between;
-		cursor: grab;
-		min-width: 1200px;
-	}
-`
 export const SmallMainContainer = styled.div`
 	display: block;
 	max-width: 977px;
@@ -103,22 +68,6 @@ export const MainContentContainer = styled(Col)`
 	@media (max-width: ${breakpoints.semixxl}px) {
 		width: 100%;
 	}
-`
-
-export const RowSkeleton = styled(Skeleton)`
-	margin-bottom: 16px;
-	.ant-skeleton-content {
-		padding: 10px 40px;
-		height: 80px;
-		background: ${({ theme }) => theme['color-base-surface-secondary']};
-		margin: 16px 0 16px 0;
-		border-radius: 12px;
-		h3,
-		ul li {
-			&::after {
-				background: linear-gradient(90deg, rgba(255, 255, 255, 0.05) 25%, rgba(0, 0, 0, 0.2) 37%, rgba(255, 255, 255, 0.05) 63%) !important;
-			}
-		}
 `
 
 export const Status = styled.div<{ visible: boolean; status?: WALLET_TICKETS }>`
